core/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-crispy-forms==1.14.0
django-jalali==6.0.0
jdatetime==4.1.0
mysqlclient==2.1.1
numpy==1.23.0
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
